Each model has a scope which returns the
ontologies which have, at least one, entry
from each model, e.g. all ontologies which are
at least in one project or have one license model,
also each model has a method which gives the
name of the instance and the number of
ontologies which have this attribute, e.g.
all ontologies which are in this project XY.